Cleanup of the previous dep parser and manifest code.
